Data of competition results of decathlon for men

The decathlon for men is a combined event in athletics consisting of 10 track and field events: 100 metres, 400 metres, 1500 metres, 110 metre hurdles, long jump, high jump, pole vault, discus throw, javelin throw, and shotput.

Usage

decathlon

Format

A tibble with 7,958 rows and 10 variables:

race100m

dbl result of 100m race in seconds

longjump

dbl result of long jump in meters

shotput

dbl result of shot put in meters

highjump

dbl result of high jump in meters

race400m

dbl result of 400m race in seconds

race110mhurdles

dbl result of 110m hurdles race in seconds

discus

dbl result of discus throw in meters

polevault

dbl result of pole valut in meters

javelinthrow

dbl result of javelin throw in meters

race1500m

dbl result of 1500m race in seconds